Add arch module & refactor some package names
