+++ arch/build.gradle
@@ -4,38 +4,45 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'com.github.dcendents.android-maven'
 group = 'com.github.paperkite'
 
-
 android {
 	compileSdkVersion 28
+	
 	defaultConfig {
 		minSdkVersion 19
 		targetSdkVersion 28
 		versionCode 1
 		versionName "1.0"
+		
 		testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+		
 	}
+	
 	buildTypes {
 		release {
 			minifyEnabled false
 			proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
 		}
 	}
-	dataBinding {
-		enabled = true
-	}
+	
 }
 
 ext {
-	version_glide = "4.4.0"
+	version_arch_lifecycle = '1.1.1'
+	version_arch_room = '1.1.1'
 }
 
 dependencies {
 	implementation fileTree(dir: 'libs', include: ['*.jar'])
+	
 	implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 	implementation "com.android.support:appcompat-v7:$version_supportlib"
 	
-	implementation "com.github.bumptech.glide:glide:$version_glide"
-	kapt "com.github.bumptech.glide:compiler:$version_glide"
+	implementation "android.arch.lifecycle:extensions:$version_arch_lifecycle"
+	kapt "android.arch.lifecycle:compiler:$version_arch_lifecycle"
+	implementation "android.arch.persistence.room:runtime:$version_arch_room"
+	kapt "android.arch.persistence.room:compiler:$version_arch_room"
+	implementation "android.arch.persistence.room:rxjava2:$version_arch_room"
+	implementation "android.arch.paging:runtime:1.0.1"
 	
 	testImplementation 'junit:junit:4.12'
