+++ arch/src/test/java/nz/co/paperkite/androidutils/arch/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package nz.co.paperkite.androidutils;
+package nz.co.paperkite.androidutils.arch;
 
 import org.junit.Test;
