+++ build.gradle
@@ -2,6 +2,8 @@
 
 buildscript {
 	ext.kotlin_version = '1.2.70'
+    ext.version_supportlib = '27.1.0'
+    
     repositories {
         google()
         jcenter()
