+++ rx/src/test/java/nz/co/paperkite/androidutils/rx/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package nz.co.paperkite.androidutils;
+package nz.co.paperkite.androidutils.rx;
 
 import org.junit.Test;
