+++ views/build.gradle
@@ -30,7 +30,7 @@ dependencies {
 	implementation fileTree(dir: 'libs', include: ['*.jar'])
 	
 	implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-	implementation 'com.android.support:appcompat-v7:28.0.0-rc02'
+	implementation "com.android.support:appcompat-v7:$version_supportlib"
 	
 	testImplementation 'junit:junit:4.12'
